requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
speedtest-cli>=2.1.3
rich>=13.7.0
toml>=0.10.2
click>=8.1.7
rapidfuzz>=3.6.0
platformdirs>=4.1.0
